fix #2799: The summary screen is shown on the mobile app when there are no alert responses and score ID to show
